refactor: continuaçao da refatoraçao dos botoes
@inglyd @LuizHenrique000

32
