+++ src/components/buttons/action/index.js
@@ -3,7 +3,7 @@ import './action.css'
 
 const ActionButton = ({ icon, onClick, text, type, disabled = false }) => {
   return (
-    <Button icon={icon} onClick={onClick} text={text} type={type} classe="button-action" disabled={disabled} />
+    <Button icon={icon} onClick={onClick} text={text} type={type} classe="button action" disabled={disabled} />
   )
 }
