+++ src/components/buttons/button/button.css
@@ -2,7 +2,6 @@
   border: 1px solid #d3d3d3;
   border-radius: 5px;
   padding: 10px 25px;
-  font-family: "Open Sans";
   font-weight: 300px;
   font-size: 16px;
 }
@@ -12,7 +11,12 @@
   opacity: 0.7;
 }
 
-.button-new-process {
+.button:disabled {
+  opacity: 0.4;
+  cursor: no-drop;
+}
+
+.new-process {
   background-color: #4444ff;
   color: #fff;
 }
@@ -28,32 +32,32 @@
   transition: 3ms;
 }
 
-.button-import,
-.button-export,
-.button-edit,
-.button-expend,
-.button-delete,
-.button-close,
+.import,
+.export,
+.edit,
+.expend,
+.delete,
+.close,
 .button-filter {
   background-color: #fff;
 }
 
-.button-edit,
-.button-expend {
+.edit,
+.expend {
   color: #003cff;
 }
 
-.button-delete {
+.delete {
   color: #ff3d00;
 }
 
-.button-close {
+.close {
   color: #aea9bd;
   border: none;
   cursor: pointer;
 }
 
-.button-submit {
+.submit {
   color: #fff;
   background-color: #003cff;
 }
