+++ src/components/buttons/default/index.js
@@ -3,7 +3,7 @@ import './default.css'
 
 const DefaultButton = ({ icon, onClick, text, type }) => {
   return (
-    < Button icon={icon} onClick={onClick} text={text} type={type} classe="button-default" />
+    <Button icon={icon} onClick={onClick} text={text} type={type} classe="button default" />
   )
 }
