+++ src/components/buttons/primary/index.js
@@ -3,7 +3,7 @@ import './primary.css'
 
 const PrimaryButton = ({ icon, onClick, text, type, disabled = false }) => {
   return (
-    < Button icon={icon} onClick={onClick} text={text} type={type} classe='button-primary' disabled={disabled} />
+    <Button icon={icon} onClick={onClick} text={text} type={type} classe='button primary' disabled={disabled} />
   )
 }
