+++ src/components/buttons/primary/primary.css
@@ -1,14 +1,10 @@
-.button-primary {
+.primary {
   background-color: rgb(68, 68, 255);
   color: white;
   border: 1px solid white;
-  border-radius: 5px;
-  font-weight: 300;
-  padding: 10px 25px;
-  cursor: pointer;
 }
 
-.button-primary:disabled {
+.primary:disabled {
   opacity: 0.3;
   cursor: not-allowed;
 }
