+++ src/components/buttons/sucess/index.js
@@ -3,7 +3,7 @@ import './sucess.css'
 
 const SucessButton = ({ icon, onClick, text, type }) => {
   return (
-    < Button icon={icon} onClick={onClick} text={text} type={type} classe="button-sucess" />
+    <Button icon={icon} onClick={onClick} text={text} type={type} classe="button sucess" />
   )
 }
