+++ src/components/buttons/sucess/sucess.css
@@ -1,13 +1,4 @@
-.button-sucess {
+.sucess {
   background-color: darkgreen;
   color: white;
-  border: 1px solid white;
-  border-radius: 5px;
-  font-weight: 300;
-  padding: 10px 25px;
-}
-
-.button-sucess:hover {
-  font-weight: bold;
-  cursor: pointer;
 }
