+++ src/pages/evaluation/style.css
@@ -45,7 +45,7 @@
   justify-content: right;
   margin-top: 20px;
 }
-a.button-default {
+a.default {
   background-color: #dddcdc;
   color: black;
   margin: 1%;
