+++ src/pages/exercises/components/toggle-row/index.js
@@ -2,7 +2,7 @@ import { faAngleDown, faPen } from '@fortawesome/free-solid-svg-icons'
 import { useState } from 'react'
 import './style.css'
 import ActionButton from '../../../../components/buttons/action'
-import Button from '../../../../components/button'
+import Button from '../../../../components/buttons/button'
 import { Status } from '../../../../components/status'
 import { client } from '../../../../service'
 import { useNavigate } from 'react-router-dom'
